+++ kalman.hpp
@@ -51,15 +51,13 @@
 #define COSDELTAMAX 0.4226183F  // cos of max +ve geomagnetic inclination angle: here 65.0 deg
 
 namespace filter::kalman {
-    void fInit_6DOF_GY_KALMAN(struct ::filter::tasks::SV_6DOF_GY_KALMAN* pthisSV,
-                              int16 iSensorFS,
-                              int16 iOverSampleRatio);
+    void fInit_6DOF_GY_KALMAN(struct ::filter::tasks::SV_6DOF_GY_KALMAN* pthisSV, int iSensorFS, int iOverSampleRatio);
     void fRun_6DOF_GY_KALMAN(struct ::filter::tasks::SV_6DOF_GY_KALMAN* pthisSV,
                              //  struct AccelSensor* pthisAccel,
                              double accel_reading[3],
                              //  struct GyroSensor* pthisGyro,
                              double gyro_reading[3],
-                             int16 ithisCoordSystem,
-                             int16 iOverSampleRatio);
+                             int ithisCoordSystem,
+                             int iOverSampleRatio);
 }  // namespace filter::kalman
 #endif  // KALMAN_HPP
